chore: 🤖 Added jest client for some convinience, few more tests
Added a test runner client in the .build/jest folder, in same manner as
bkg - test-runner. We wanna utilize some of the jest functionalities, in
a more structured manner. Added some dummy tests for FormField.
