+++ .eslintrc.json
@@ -182,7 +182,7 @@
         "func-names": "error",
         "func-style": "off",
         "function-call-argument-newline": ["error", "consistent"],
-        "function-paren-newline": "error",
+        "function-paren-newline": "off",
         "id-blacklist": "error",
         "id-length": "error",
         "id-match": "error",
