--- setupTestEnv.ts
@@ -1,7 +1,0 @@
-// @ts-nocheck
-import '@testing-library/jest-dom/extend-expect';
-import "@testing-library/jest-dom";
-import Enzyme from 'enzyme';
-import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
-
-Enzyme.configure({ adapter: new Adapter() });
